+++ integration/simple-long/import_dir/thing.ts
@@ -1,6 +1,6 @@
 import { Timestamp } from '../google/protobuf/timestamp';
 import * as Long from 'long';
-import { Writer, Reader } from 'protobufjs/minimal';
+import { Writer, Reader, util, configure } from 'protobufjs/minimal';
 
 
 export interface ImportedThing {
@@ -80,10 +80,15 @@ export const ImportedThing = {
   },
   toJSON(message: ImportedThing): unknown {
     const obj: any = {};
-    obj.createdAt = message.createdAt !== undefined ? message.createdAt.toISOString() : null;
+    message.createdAt !== undefined && (obj.createdAt = message.createdAt !== undefined ? message.createdAt.toISOString() : null);
     return obj;
   },
 };
+
+if (util.Long !== Long as any) {
+  util.Long = Long as any;
+  configure();
+}
 
 type Builtin = Date | Function | Uint8Array | string | number | undefined;
 type DeepPartial<T> = T extends Builtin