+++ integration/simple/simple.ts
@@ -3,7 +3,7 @@
 //
 import { ImportedThing } from './import_dir/thing';
 import { DateMessage } from './google/type/date';
-import { Reader, Writer } from 'protobufjs/minimal';
+import { Reader, Writer, util, configure } from 'protobufjs/minimal';
 import { Timestamp } from './google/protobuf/timestamp';
 import * as Long from 'long';
 import { StringValue, Int32Value, BoolValue } from './google/protobuf/wrappers';
@@ -35,6 +35,7 @@ export interface Simple {
   thing: ImportedThing | undefined;
   blobs: Uint8Array[];
   birthday: DateMessage | undefined;
+  blob: Uint8Array;
 }
 
 export interface Child {
@@ -119,6 +120,15 @@ export interface SimpleWithSnakeCaseMap_EntitiesByIdEntry {
   value: Entity | undefined;
 }
 
+export interface SimpleWithMapOfEnums {
+  enumsById: { [key: number]: StateEnum };
+}
+
+export interface SimpleWithMapOfEnums_EnumsByIdEntry {
+  key: number;
+  value: StateEnum;
+}
+
 export interface PingRequest {
   input: string;
 }
@@ -140,6 +150,29 @@ export interface Numbers {
   fixed64: number;
   sfixed32: number;
   sfixed64: number;
+}
+
+/**
+ * * For testing proto3's field presence feature.  */
+export interface SimpleButOptional {
+  /**
+   *  Name field
+   */
+  name?: string | undefined;
+  /**
+   *  Age  */
+  age?: number | undefined;
+  /**
+   *  This comment will also attach
+   */
+  createdAt?: Date | undefined;
+  child?: Child | undefined;
+  state?: StateEnum | undefined;
+  /**
+   *  A thing (imported from thing)
+   */
+  thing?: ImportedThing | undefined;
+  birthday?: DateMessage | undefined;
 }
 
 export interface Empty {
@@ -214,6 +247,14 @@ const baseSimpleWithSnakeCaseMap_EntitiesByIdEntry: object = {
   key: 0,
 };
 
+const baseSimpleWithMapOfEnums: object = {
+};
+
+const baseSimpleWithMapOfEnums_EnumsByIdEntry: object = {
+  key: 0,
+  value: 0,
+};
+
 const basePingRequest: object = {
   input: "",
 };
@@ -235,6 +276,9 @@ const baseNumbers: object = {
   fixed64: 0,
   sfixed32: 0,
   sfixed64: 0,
+};
+
+const baseSimpleButOptional: object = {
 };
 
 const baseEmpty: object = {
@@ -447,6 +491,7 @@ export const Simple = {
     if (message.birthday !== undefined && message.birthday !== undefined) {
       DateMessage.encode(message.birthday, writer.uint32(98).fork()).ldelim();
     }
+    writer.uint32(106).bytes(message.blob);
     return writer;
   },
   decode(input: Uint8Array | Reader, length?: number): Simple {
@@ -510,6 +555,9 @@ export const Simple = {
           break;
         case 12:
           message.birthday = DateMessage.decode(reader, reader.uint32());
+          break;
+        case 13:
+          message.blob = reader.bytes();
           break;
         default:
           reader.skipType(tag & 7);
@@ -585,6 +633,9 @@ export const Simple = {
     } else {
       message.birthday = undefined;
     }
+    if (object.blob !== undefined && object.blob !== null) {
+      message.blob = bytesFromBase64(object.blob);
+    }
     return message;
   },
   fromPartial(object: DeepPartial<Simple>): Simple {
@@ -654,27 +705,32 @@ export const Simple = {
     } else {
       message.birthday = undefined;
     }
+    if (object.blob !== undefined && object.blob !== null) {
+      message.blob = object.blob;
+    } else {
+      message.blob = new Uint8Array();
+    }
     return message;
   },
   toJSON(message: Simple): unknown {
     const obj: any = {};
-    obj.name = message.name || "";
-    obj.age = message.age || 0;
-    obj.createdAt = message.createdAt !== undefined ? message.createdAt.toISOString() : null;
-    obj.child = message.child ? Child.toJSON(message.child) : undefined;
-    obj.state = stateEnumToJSON(message.state);
+    message.name !== undefined && (obj.name = message.name);
+    message.age !== undefined && (obj.age = message.age);
+    message.createdAt !== undefined && (obj.createdAt = message.createdAt !== undefined ? message.createdAt.toISOString() : null);
+    message.child !== undefined && (obj.child = message.child ? Child.toJSON(message.child) : undefined);
+    message.state !== undefined && (obj.state = stateEnumToJSON(message.state));
     if (message.grandChildren) {
       obj.grandChildren = message.grandChildren.map(e => e ? Child.toJSON(e) : undefined);
     } else {
       obj.grandChildren = [];
     }
     if (message.coins) {
-      obj.coins = message.coins.map(e => e || 0);
+      obj.coins = message.coins.map(e => e);
     } else {
       obj.coins = [];
     }
     if (message.snacks) {
-      obj.snacks = message.snacks.map(e => e || "");
+      obj.snacks = message.snacks.map(e => e);
     } else {
       obj.snacks = [];
     }
@@ -683,13 +739,14 @@ export const Simple = {
     } else {
       obj.oldStates = [];
     }
-    obj.thing = message.thing ? ImportedThing.toJSON(message.thing) : undefined;
+    message.thing !== undefined && (obj.thing = message.thing ? ImportedThing.toJSON(message.thing) : undefined);
     if (message.blobs) {
-      obj.blobs = message.blobs.map(e => e !== undefined ? base64FromBytes(e) : undefined);
+      obj.blobs = message.blobs.map(e => base64FromBytes(e !== undefined ? e : new Uint8Array()));
     } else {
       obj.blobs = [];
     }
-    obj.birthday = message.birthday ? DateMessage.toJSON(message.birthday) : undefined;
+    message.birthday !== undefined && (obj.birthday = message.birthday ? DateMessage.toJSON(message.birthday) : undefined);
+    message.blob !== undefined && (obj.blob = base64FromBytes(message.blob !== undefined ? message.blob : new Uint8Array()));
     return obj;
   },
 };
@@ -750,8 +807,8 @@ export const Child = {
   },
   toJSON(message: Child): unknown {
     const obj: any = {};
-    obj.name = message.name || "";
-    obj.type = child_TypeToJSON(message.type);
+    message.name !== undefined && (obj.name = message.name);
+    message.type !== undefined && (obj.type = child_TypeToJSON(message.type));
     return obj;
   },
 };
@@ -828,9 +885,9 @@ export const Nested = {
   },
   toJSON(message: Nested): unknown {
     const obj: any = {};
-    obj.name = message.name || "";
-    obj.message = message.message ? Nested_InnerMessage.toJSON(message.message) : undefined;
-    obj.state = nested_InnerEnumToJSON(message.state);
+    message.name !== undefined && (obj.name = message.name);
+    message.message !== undefined && (obj.message = message.message ? Nested_InnerMessage.toJSON(message.message) : undefined);
+    message.state !== undefined && (obj.state = nested_InnerEnumToJSON(message.state));
     return obj;
   },
 };
@@ -893,8 +950,8 @@ export const Nested_InnerMessage = {
   },
   toJSON(message: Nested_InnerMessage): unknown {
     const obj: any = {};
-    obj.name = message.name || "";
-    obj.deep = message.deep ? Nested_InnerMessage_DeepMessage.toJSON(message.deep) : undefined;
+    message.name !== undefined && (obj.name = message.name);
+    message.deep !== undefined && (obj.deep = message.deep ? Nested_InnerMessage_DeepMessage.toJSON(message.deep) : undefined);
     return obj;
   },
 };
@@ -941,17 +998,17 @@ export const Nested_InnerMessage_DeepMessage = {
   },
   toJSON(message: Nested_InnerMessage_DeepMessage): unknown {
     const obj: any = {};
-    obj.name = message.name || "";
+    message.name !== undefined && (obj.name = message.name);
     return obj;
   },
 };
 
 export const OneOfMessage = {
   encode(message: OneOfMessage, writer: Writer = Writer.create()): Writer {
-    if (message.first !== undefined && message.first !== "") {
+    if (message.first !== undefined) {
       writer.uint32(10).string(message.first);
     }
-    if (message.last !== undefined && message.last !== "") {
+    if (message.last !== undefined) {
       writer.uint32(18).string(message.last);
     }
     return writer;
@@ -1006,8 +1063,8 @@ export const OneOfMessage = {
   },
   toJSON(message: OneOfMessage): unknown {
     const obj: any = {};
-    obj.first = message.first || undefined;
-    obj.last = message.last || undefined;
+    message.first !== undefined && (obj.first = message.first);
+    message.last !== undefined && (obj.last = message.last);
     return obj;
   },
 };
@@ -1126,16 +1183,16 @@ export const SimpleWithWrappers = {
   },
   toJSON(message: SimpleWithWrappers): unknown {
     const obj: any = {};
-    obj.name = message.name || undefined;
-    obj.age = message.age || undefined;
-    obj.enabled = message.enabled || undefined;
+    message.name !== undefined && (obj.name = message.name);
+    message.age !== undefined && (obj.age = message.age);
+    message.enabled !== undefined && (obj.enabled = message.enabled);
     if (message.coins) {
-      obj.coins = message.coins.map(e => e || undefined);
+      obj.coins = message.coins.map(e => e);
     } else {
       obj.coins = [];
     }
     if (message.snacks) {
-      obj.snacks = message.snacks.map(e => e || undefined);
+      obj.snacks = message.snacks.map(e => e);
     } else {
       obj.snacks = [];
     }
@@ -1185,7 +1242,7 @@ export const Entity = {
   },
   toJSON(message: Entity): unknown {
     const obj: any = {};
-    obj.id = message.id || 0;
+    message.id !== undefined && (obj.id = message.id);
     return obj;
   },
 };
@@ -1430,8 +1487,8 @@ export const SimpleWithMap_EntitiesByIdEntry = {
   },
   toJSON(message: SimpleWithMap_EntitiesByIdEntry): unknown {
     const obj: any = {};
-    obj.key = message.key || 0;
-    obj.value = message.value ? Entity.toJSON(message.value) : undefined;
+    message.key !== undefined && (obj.key = message.key);
+    message.value !== undefined && (obj.value = message.value ? Entity.toJSON(message.value) : undefined);
     return obj;
   },
 };
@@ -1492,8 +1549,8 @@ export const SimpleWithMap_NameLookupEntry = {
   },
   toJSON(message: SimpleWithMap_NameLookupEntry): unknown {
     const obj: any = {};
-    obj.key = message.key || "";
-    obj.value = message.value || "";
+    message.key !== undefined && (obj.key = message.key);
+    message.value !== undefined && (obj.value = message.value);
     return obj;
   },
 };
@@ -1554,8 +1611,8 @@ export const SimpleWithMap_IntLookupEntry = {
   },
   toJSON(message: SimpleWithMap_IntLookupEntry): unknown {
     const obj: any = {};
-    obj.key = message.key || 0;
-    obj.value = message.value || 0;
+    message.key !== undefined && (obj.key = message.key);
+    message.value !== undefined && (obj.value = message.value);
     return obj;
   },
 };
@@ -1618,8 +1675,8 @@ export const SimpleWithMap_MapOfTimestampsEntry = {
   },
   toJSON(message: SimpleWithMap_MapOfTimestampsEntry): unknown {
     const obj: any = {};
-    obj.key = message.key || "";
-    obj.value = message.value !== undefined ? message.value.toISOString() : null;
+    message.key !== undefined && (obj.key = message.key);
+    message.value !== undefined && (obj.value = message.value !== undefined ? message.value.toISOString() : null);
     return obj;
   },
 };
@@ -1671,13 +1728,15 @@ export const SimpleWithMap_MapOfBytesEntry = {
     }
     if (object.value !== undefined && object.value !== null) {
       message.value = object.value;
+    } else {
+      message.value = new Uint8Array();
     }
     return message;
   },
   toJSON(message: SimpleWithMap_MapOfBytesEntry): unknown {
     const obj: any = {};
-    obj.key = message.key || "";
-    obj.value = message.value !== undefined ? base64FromBytes(message.value) : undefined;
+    message.key !== undefined && (obj.key = message.key);
+    message.value !== undefined && (obj.value = base64FromBytes(message.value !== undefined ? message.value : new Uint8Array()));
     return obj;
   },
 };
@@ -1802,8 +1861,132 @@ export const SimpleWithSnakeCaseMap_EntitiesByIdEntry = {
   },
   toJSON(message: SimpleWithSnakeCaseMap_EntitiesByIdEntry): unknown {
     const obj: any = {};
-    obj.key = message.key || 0;
-    obj.value = message.value ? Entity.toJSON(message.value) : undefined;
+    message.key !== undefined && (obj.key = message.key);
+    message.value !== undefined && (obj.value = message.value ? Entity.toJSON(message.value) : undefined);
+    return obj;
+  },
+};
+
+export const SimpleWithMapOfEnums = {
+  encode(message: SimpleWithMapOfEnums, writer: Writer = Writer.create()): Writer {
+    Object.entries(message.enumsById).forEach(([key, value]) => {
+      SimpleWithMapOfEnums_EnumsByIdEntry.encode({ key: key as any, value }, writer.uint32(10).fork()).ldelim();
+    })
+    return writer;
+  },
+  decode(input: Uint8Array | Reader, length?: number): SimpleWithMapOfEnums {
+    const reader = input instanceof Uint8Array ? new Reader(input) : input;
+    let end = length === undefined ? reader.len : reader.pos + length;
+    const message = { ...baseSimpleWithMapOfEnums } as SimpleWithMapOfEnums;
+    message.enumsById = {};
+    while (reader.pos < end) {
+      const tag = reader.uint32();
+      switch (tag >>> 3) {
+        case 1:
+          const entry1 = SimpleWithMapOfEnums_EnumsByIdEntry.decode(reader, reader.uint32());
+          if (entry1.value !== undefined) {
+            message.enumsById[entry1.key] = entry1.value;
+          }
+          break;
+        default:
+          reader.skipType(tag & 7);
+          break;
+      }
+    }
+    return message;
+  },
+  fromJSON(object: any): SimpleWithMapOfEnums {
+    const message = { ...baseSimpleWithMapOfEnums } as SimpleWithMapOfEnums;
+    message.enumsById = {};
+    if (object.enumsById !== undefined && object.enumsById !== null) {
+      Object.entries(object.enumsById).forEach(([key, value]) => {
+        message.enumsById[Number(key)] = value as number;
+      })
+    }
+    return message;
+  },
+  fromPartial(object: DeepPartial<SimpleWithMapOfEnums>): SimpleWithMapOfEnums {
+    const message = { ...baseSimpleWithMapOfEnums } as SimpleWithMapOfEnums;
+    message.enumsById = {};
+    if (object.enumsById !== undefined && object.enumsById !== null) {
+      Object.entries(object.enumsById).forEach(([key, value]) => {
+        if (value !== undefined) {
+          message.enumsById[Number(key)] = value as number;
+        }
+      })
+    }
+    return message;
+  },
+  toJSON(message: SimpleWithMapOfEnums): unknown {
+    const obj: any = {};
+    obj.enumsById = {};
+    if (message.enumsById) {
+      Object.entries(message.enumsById).forEach(([k, v]) => {
+        obj.enumsById[k] = stateEnumToJSON(v);
+      })
+    }
+    return obj;
+  },
+};
+
+export const SimpleWithMapOfEnums_EnumsByIdEntry = {
+  encode(message: SimpleWithMapOfEnums_EnumsByIdEntry, writer: Writer = Writer.create()): Writer {
+    writer.uint32(8).int32(message.key);
+    writer.uint32(16).int32(message.value);
+    return writer;
+  },
+  decode(input: Uint8Array | Reader, length?: number): SimpleWithMapOfEnums_EnumsByIdEntry {
+    const reader = input instanceof Uint8Array ? new Reader(input) : input;
+    let end = length === undefined ? reader.len : reader.pos + length;
+    const message = { ...baseSimpleWithMapOfEnums_EnumsByIdEntry } as SimpleWithMapOfEnums_EnumsByIdEntry;
+    while (reader.pos < end) {
+      const tag = reader.uint32();
+      switch (tag >>> 3) {
+        case 1:
+          message.key = reader.int32();
+          break;
+        case 2:
+          message.value = reader.int32() as any;
+          break;
+        default:
+          reader.skipType(tag & 7);
+          break;
+      }
+    }
+    return message;
+  },
+  fromJSON(object: any): SimpleWithMapOfEnums_EnumsByIdEntry {
+    const message = { ...baseSimpleWithMapOfEnums_EnumsByIdEntry } as SimpleWithMapOfEnums_EnumsByIdEntry;
+    if (object.key !== undefined && object.key !== null) {
+      message.key = Number(object.key);
+    } else {
+      message.key = 0;
+    }
+    if (object.value !== undefined && object.value !== null) {
+      message.value = stateEnumFromJSON(object.value);
+    } else {
+      message.value = 0;
+    }
+    return message;
+  },
+  fromPartial(object: DeepPartial<SimpleWithMapOfEnums_EnumsByIdEntry>): SimpleWithMapOfEnums_EnumsByIdEntry {
+    const message = { ...baseSimpleWithMapOfEnums_EnumsByIdEntry } as SimpleWithMapOfEnums_EnumsByIdEntry;
+    if (object.key !== undefined && object.key !== null) {
+      message.key = object.key;
+    } else {
+      message.key = 0;
+    }
+    if (object.value !== undefined && object.value !== null) {
+      message.value = object.value;
+    } else {
+      message.value = 0;
+    }
+    return message;
+  },
+  toJSON(message: SimpleWithMapOfEnums_EnumsByIdEntry): unknown {
+    const obj: any = {};
+    message.key !== undefined && (obj.key = message.key);
+    message.value !== undefined && (obj.value = stateEnumToJSON(message.value));
     return obj;
   },
 };
@@ -1850,7 +2033,7 @@ export const PingRequest = {
   },
   toJSON(message: PingRequest): unknown {
     const obj: any = {};
-    obj.input = message.input || "";
+    message.input !== undefined && (obj.input = message.input);
     return obj;
   },
 };
@@ -1897,7 +2080,7 @@ export const PingResponse = {
   },
   toJSON(message: PingResponse): unknown {
     const obj: any = {};
-    obj.output = message.output || "";
+    message.output !== undefined && (obj.output = message.output);
     return obj;
   },
 };
@@ -2098,18 +2281,169 @@ export const Numbers = {
   },
   toJSON(message: Numbers): unknown {
     const obj: any = {};
-    obj.double = message.double || 0;
-    obj.float = message.float || 0;
-    obj.int32 = message.int32 || 0;
-    obj.int64 = message.int64 || 0;
-    obj.uint32 = message.uint32 || 0;
-    obj.uint64 = message.uint64 || 0;
-    obj.sint32 = message.sint32 || 0;
-    obj.sint64 = message.sint64 || 0;
-    obj.fixed32 = message.fixed32 || 0;
-    obj.fixed64 = message.fixed64 || 0;
-    obj.sfixed32 = message.sfixed32 || 0;
-    obj.sfixed64 = message.sfixed64 || 0;
+    message.double !== undefined && (obj.double = message.double);
+    message.float !== undefined && (obj.float = message.float);
+    message.int32 !== undefined && (obj.int32 = message.int32);
+    message.int64 !== undefined && (obj.int64 = message.int64);
+    message.uint32 !== undefined && (obj.uint32 = message.uint32);
+    message.uint64 !== undefined && (obj.uint64 = message.uint64);
+    message.sint32 !== undefined && (obj.sint32 = message.sint32);
+    message.sint64 !== undefined && (obj.sint64 = message.sint64);
+    message.fixed32 !== undefined && (obj.fixed32 = message.fixed32);
+    message.fixed64 !== undefined && (obj.fixed64 = message.fixed64);
+    message.sfixed32 !== undefined && (obj.sfixed32 = message.sfixed32);
+    message.sfixed64 !== undefined && (obj.sfixed64 = message.sfixed64);
+    return obj;
+  },
+};
+
+export const SimpleButOptional = {
+  encode(message: SimpleButOptional, writer: Writer = Writer.create()): Writer {
+    if (message.name !== undefined) {
+      writer.uint32(10).string(message.name);
+    }
+    if (message.age !== undefined) {
+      writer.uint32(16).int32(message.age);
+    }
+    if (message.createdAt !== undefined) {
+      Timestamp.encode(toTimestamp(message.createdAt), writer.uint32(74).fork()).ldelim();
+    }
+    if (message.child !== undefined) {
+      Child.encode(message.child, writer.uint32(26).fork()).ldelim();
+    }
+    if (message.state !== undefined) {
+      writer.uint32(32).int32(message.state);
+    }
+    if (message.thing !== undefined) {
+      ImportedThing.encode(message.thing, writer.uint32(82).fork()).ldelim();
+    }
+    if (message.birthday !== undefined) {
+      DateMessage.encode(message.birthday, writer.uint32(98).fork()).ldelim();
+    }
+    return writer;
+  },
+  decode(input: Uint8Array | Reader, length?: number): SimpleButOptional {
+    const reader = input instanceof Uint8Array ? new Reader(input) : input;
+    let end = length === undefined ? reader.len : reader.pos + length;
+    const message = { ...baseSimpleButOptional } as SimpleButOptional;
+    while (reader.pos < end) {
+      const tag = reader.uint32();
+      switch (tag >>> 3) {
+        case 1:
+          message.name = reader.string();
+          break;
+        case 2:
+          message.age = reader.int32();
+          break;
+        case 9:
+          message.createdAt = fromTimestamp(Timestamp.decode(reader, reader.uint32()));
+          break;
+        case 3:
+          message.child = Child.decode(reader, reader.uint32());
+          break;
+        case 4:
+          message.state = reader.int32() as any;
+          break;
+        case 10:
+          message.thing = ImportedThing.decode(reader, reader.uint32());
+          break;
+        case 12:
+          message.birthday = DateMessage.decode(reader, reader.uint32());
+          break;
+        default:
+          reader.skipType(tag & 7);
+          break;
+      }
+    }
+    return message;
+  },
+  fromJSON(object: any): SimpleButOptional {
+    const message = { ...baseSimpleButOptional } as SimpleButOptional;
+    if (object.name !== undefined && object.name !== null) {
+      message.name = String(object.name);
+    } else {
+      message.name = undefined;
+    }
+    if (object.age !== undefined && object.age !== null) {
+      message.age = Number(object.age);
+    } else {
+      message.age = undefined;
+    }
+    if (object.createdAt !== undefined && object.createdAt !== null) {
+      message.createdAt = fromJsonTimestamp(object.createdAt);
+    } else {
+      message.createdAt = undefined;
+    }
+    if (object.child !== undefined && object.child !== null) {
+      message.child = Child.fromJSON(object.child);
+    } else {
+      message.child = undefined;
+    }
+    if (object.state !== undefined && object.state !== null) {
+      message.state = stateEnumFromJSON(object.state);
+    } else {
+      message.state = undefined;
+    }
+    if (object.thing !== undefined && object.thing !== null) {
+      message.thing = ImportedThing.fromJSON(object.thing);
+    } else {
+      message.thing = undefined;
+    }
+    if (object.birthday !== undefined && object.birthday !== null) {
+      message.birthday = DateMessage.fromJSON(object.birthday);
+    } else {
+      message.birthday = undefined;
+    }
+    return message;
+  },
+  fromPartial(object: DeepPartial<SimpleButOptional>): SimpleButOptional {
+    const message = { ...baseSimpleButOptional } as SimpleButOptional;
+    if (object.name !== undefined && object.name !== null) {
+      message.name = object.name;
+    } else {
+      message.name = undefined;
+    }
+    if (object.age !== undefined && object.age !== null) {
+      message.age = object.age;
+    } else {
+      message.age = undefined;
+    }
+    if (object.createdAt !== undefined && object.createdAt !== null) {
+      message.createdAt = object.createdAt;
+    } else {
+      message.createdAt = undefined;
+    }
+    if (object.child !== undefined && object.child !== null) {
+      message.child = Child.fromPartial(object.child);
+    } else {
+      message.child = undefined;
+    }
+    if (object.state !== undefined && object.state !== null) {
+      message.state = object.state;
+    } else {
+      message.state = undefined;
+    }
+    if (object.thing !== undefined && object.thing !== null) {
+      message.thing = ImportedThing.fromPartial(object.thing);
+    } else {
+      message.thing = undefined;
+    }
+    if (object.birthday !== undefined && object.birthday !== null) {
+      message.birthday = DateMessage.fromPartial(object.birthday);
+    } else {
+      message.birthday = undefined;
+    }
+    return message;
+  },
+  toJSON(message: SimpleButOptional): unknown {
+    const obj: any = {};
+    message.name !== undefined && (obj.name = message.name);
+    message.age !== undefined && (obj.age = message.age);
+    message.createdAt !== undefined && (obj.createdAt = message.createdAt !== undefined ? message.createdAt.toISOString() : null);
+    message.child !== undefined && (obj.child = message.child ? Child.toJSON(message.child) : undefined);
+    message.state !== undefined && (obj.state = message.state !== undefined ? stateEnumToJSON(message.state) : undefined);
+    message.thing !== undefined && (obj.thing = message.thing ? ImportedThing.toJSON(message.thing) : undefined);
+    message.birthday !== undefined && (obj.birthday = message.birthday ? DateMessage.toJSON(message.birthday) : undefined);
     return obj;
   },
 };
@@ -2145,6 +2479,11 @@ export const Empty = {
     return obj;
   },
 };
+
+if (util.Long !== Long as any) {
+  util.Long = Long as any;
+  configure();
+}
 
 interface WindowBase64 {
   atob(b64: string): string;